HDX-9781 Move org view as table
